+++ b_aws_testing_framework/tools/cdk_testing/cdk_tool_config.py
@@ -1,3 +1,6 @@
+from typing import Optional
+
+
 class CdkToolConfig:
     """
     Configuration class for aws cdk tool.
@@ -6,7 +9,8 @@ class CdkToolConfig:
     def __init__(
             self,
             cdk_app_path: str,
-            destroy_before_preparing: bool = True
+            destroy_before_preparing: bool = True,
+            project_root_path: Optional[str] = None
     ) -> None:
         """
         Constructor.
@@ -15,9 +19,12 @@ class CdkToolConfig:
         :param destroy_before_preparing: A flag indicating whether the infrastructure manager should attempt
         to clean (destroy) the infrastructure before creating one. This is useful in cases there are some
         resource leftovers from a previous run.
+        :param project_root_path: An absolute path to the root of your whole project. This variable is used
+        to add to python path. If you face any import errors - it might be a good idea to set this variable.
         """
         self.__cdk_app_path = cdk_app_path
         self.__destroy_before_preparing = destroy_before_preparing
+        self.__project_root_path = project_root_path
 
     @property
     def cdk_app_path(self):
@@ -26,3 +33,7 @@ class CdkToolConfig:
     @property
     def destroy_before_preparing(self):
         return self.__destroy_before_preparing
+
+    @property
+    def project_root_path(self):
+        return self.__project_root_path