+++ b_aws_testing_framework_test/tools/testing_with_cdk/conftest.py
@@ -1,5 +1,7 @@
 import logging
-import os
+
+from os.path import dirname as dirn
+from os.path import abspath as apat
 
 from b_aws_testing_framework.credentials import Credentials
 from b_aws_testing_framework.tools.cdk_testing.cdk_tool_config import CdkToolConfig
@@ -7,18 +9,25 @@ from b_aws_testing_framework.tools.cdk_testing.testing_manager import TestingMan
 
 logging.basicConfig(format='%(levelname)s:%(message)s', level=logging.INFO)
 
-CDK_PATH = f'{os.path.dirname(os.path.abspath(__file__))}'
+CDK_PATH = dirn(apat(__file__))
+ROOT_PATH = dirn(dirn(dirn(dirn(apat(__file__)))))
 
 
 def pytest_configure(*args, **kwargs):
     """
     Called after command line options have been parsed and all plugins and initial conftest files been loaded.
     """
-    TestingManager(Credentials(), CdkToolConfig(CDK_PATH)).prepare_infrastructure()
+    TestingManager(Credentials(), CdkToolConfig(
+        cdk_app_path=CDK_PATH,
+        project_root_path=ROOT_PATH
+    )).prepare_infrastructure()
 
 
 def pytest_unconfigure(*args, **kwargs):
     """
     Called before test process is exited.
     """
-    TestingManager(Credentials(), CdkToolConfig(CDK_PATH)).destroy_infrastructure()
+    TestingManager(Credentials(), CdkToolConfig(
+        cdk_app_path=CDK_PATH,
+        project_root_path=ROOT_PATH
+    )).destroy_infrastructure()